web/requirements.txt
Django==1.6.1
South==0.8.4
argparse==1.2.1
django-extensions==1.3.3
djangorestframework==2.3.12
dnspython==1.11.1
ipython==1.1.0
requests==2.2.1
six==1.5.2
wsgiref==0.1.2
